refact: move learningpath and article api to separate files
