+++ src/data/learningpathApi.ts
@@ -1,0 +1,51 @@
+/**
+ * Copyright (c) 2018-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import { fetch, resolveJson } from '../utils/apiHelpers';
+
+export async function fetchLearningpaths(
+  learningpathIds: string[],
+  context: Context,
+): Promise<GQLMeta[]> {
+  const response = await fetch(
+    `/learningpath-api/v2/learningpaths/?language=${
+      context.language
+    }&fallback=true&ids=${learningpathIds.join(',')}`,
+    context,
+  );
+  const json = await resolveJson(response);
+
+  // The api does not always return the exact number of results as ids provided.
+  // So always map over ids so that dataLoader gets the right amount of results in correct order.
+  return learningpathIds.map(id => {
+    const learningpath = json.results.find((item: { id: number }) => {
+      return item.id.toString() === id;
+    });
+
+    if (learningpath) {
+      return {
+        id: learningpath.id,
+        title: learningpath.title.title,
+        introduction: learningpath.introduction
+          ? learningpath.introduction.introduction
+          : undefined,
+        metaDescription: learningpath.description
+          ? learningpath.description.description
+          : undefined,
+        lastUpdated: learningpath.lastUpdated,
+        metaImage: {
+          url: learningpath.coverPhotoUrl,
+          alt: learningpath.introduction
+            ? learningpath.introduction.introduction
+            : '',
+        },
+      };
+    }
+    return null;
+  });
+}
